r5548/sofa-dev : FIX: duplication. Two methods were doing the same thing in Simulation, printXML and exportXML. Now, only exportXML exists.

+++ modules/sofa/simulation/common/Visitor.cpp
@@ -152,7 +152,7 @@ void Visitor::stopDumpVisitor()
 
 double Visitor::getTimeSpent(ctime_t initTime, ctime_t endTime)
 {
-    return endTime-initTime;
+    return (double)(endTime-initTime);
 }
 
 void Visitor::printNode(const std::string &type, const std::string &name, const TRACE_ARGUMENT &arguments)
